requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
pydantic>=2.0.0
python-dotenv>=1.0.0
httpx>=0.26.0
pytest>=8.0.0
PyJWT>=2.8.0
passlib[bcrypt]>=1.7.4
